distance can now be a decimal, calculations changed to decimal arithmetic

+++ app/models/workout.rb
@@ -9,10 +9,10 @@ class Workout < ActiveRecord::Base
     # all workout records within the indicated time range
     # that are associated with the current workout instance.
     def sum_field_by_time_range(field, time_f)
-        self.workout_records.select{ |r| time_f.call(r) }.map{ |r| r.send(field).to_i }.sum
+        self.workout_records.select{ |r| time_f.call(r) }.map{ |r| r.send(field).to_f }.sum
     end
 
     def sum_field_by_custom_time_range(field, time_f, start_time, end_time)
-        self.workout_records.select{ |r| time_f.call(start_time, end_time, r) }.map{ |r| r.send(field).to_i }.sum
+        self.workout_records.select{ |r| time_f.call(start_time, end_time, r) }.map{ |r| r.send(field).to_f }.sum
     end
 end
